+++ examples/lastrev-next-starter/packages/web/cypress/plugins/index.js
@@ -12,8 +12,7 @@
 // This function is called when a project is opened or re-opened (e.g. due to
 // the project's config changing)
 
-const path = require('path');
-const toPath = (_path) => path.join(process.cwd(), _path);
+const generatePages = require('./generatePages');
 
 /**
  * @type {Cypress.PluginConfig}
@@ -23,5 +22,6 @@ module.exports = (on, config) => {
   // `on` is used to hook into various events Cypress emits
   // `config` is the resolved Cypress config
   require('@cypress/code-coverage/task')(on, config);
+  generatePages();
   return config;
 };