+++ packages/component-library/cypress/plugins/index.js
@@ -46,5 +46,6 @@ module.exports = (on, config) => {
       }
     });
   }
+
   return config;
 };